Refactoring project structure: placed files in folders, deleted storyboard
