+++ Pr2503/App/AppDelegate.swift
@@ -1,0 +1,4 @@
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate { }
